+++ packages/v4-sdk/src/utils/index.ts
@@ -5,3 +5,4 @@ export * from './getAddLiquidityParameters'
 export * from './getBinIds'
 export * from './getPoolId'
 export * from './math'
+export * from './poolKey'
